% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateNumeric.R
\name{dateNumeric}
\alias{dateNumeric}
\title{Convert dates according to date format}
\usage{
dateNumeric(date, format)
}
\arguments{
\item{date}{input a data extracted from labels, character}

\item{format}{input format of the date, character}
}
\value{
Returns a date of numeric type, numeric
}
\description{
Convert dates according to date format
}
\examples{

dateNumeric(date="1999-12-07", format="yyyy-MM-dd")

}
