% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateType1.R
\name{dateType1}
\alias{dateType1}
\title{Method 1 for finding the date inside the label}
\usage{
dateType1(tree, order)
}
\arguments{
\item{tree}{A tree of sequences, phylo}

\item{order}{Location of the date, character or numeric}
}
\value{
date, character
}
\description{
Method 1 for finding the date inside the label
}
\examples{

data("MCC_FluA_H3_tree")
dateType1(MCC_FluA_H3_tree, "last")

}
