% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateType2.R
\name{dateType2}
\alias{dateType2}
\title{Method 2 for finding the date inside the label}
\usage{
dateType2(tree, order, prefix)
}
\arguments{
\item{tree}{A tree of sequences, phylo}

\item{order}{Location of the date, character or numeric}

\item{prefix}{prefix for dates, character}
}
\value{
date, character
}
\description{
Method 2 for finding the date inside the label
}
\examples{

data("MCC_FluA_H3_tree")
dateType2(MCC_FluA_H3_tree, "last","/")

}
