% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateType3.R
\name{dateType3}
\alias{dateType3}
\title{Method 3 for finding the date inside the label}
\usage{
dateType3(tree, pattern)
}
\arguments{
\item{tree}{A tree of sequences, phylo}

\item{pattern}{Canonical matching command, character}
}
\value{
date, character
}
\description{
Method 3 for finding the date inside the label
}
\examples{

data("MCC_FluA_H3_tree")
dateType3(MCC_FluA_H3_tree, "(?<=/)\\\\d+$")

}
