% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdivergence.R
\name{getdivergence}
\alias{getdivergence}
\title{Calculating the divergence of sequences}
\usage{
getdivergence(tree, date, method)
}
\arguments{
\item{tree}{A tree of sequences, phylo}

\item{date}{dates of numeric type, numeric}

\item{method}{one of "correlation", "rms", or "rsquared", character}
}
\value{
the divergence of sequences, data.frame
}
\description{
Calculating the divergence of sequences
}
\examples{

data("MCC_FluA_H3_tree")
date <- dateType3(MCC_FluA_H3_tree, "(?<=/)\\\\d+$")
date <- dateNumeric(date, "yyyy")
getdivergence(MCC_FluA_H3_tree, date, "rms") 

}
