% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCC_FluA_H3_tree}
\alias{MCC_FluA_H3_tree}
\title{Example data: a tree of 76 H3 hemagglutinin gene sequences of
 a lineage containing swine and human influenza A viruses}
\format{
a tree with 76 sequences
}
\value{
a tree, phylo
}
\description{
This example data was
reported on Liang et al. 2014
}
\examples{
data(MCC_FluA_H3_tree)
}
\keyword{data}
