% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meangroup.R
\name{meangroup}
\alias{meangroup}
\title{Combining data from the same years}
\usage{
meangroup(d)
}
\arguments{
\item{d}{a data frame with "time" in the column name}
}
\value{
The processed data frame, data.frame
}
\description{
Combining data from the same years
}
\examples{

x <- c(1999, 2002  ,2005, 2000,2004 ,2004, 1999)
y <- c(1, 1.5, 2, 3 ,4 ,5 ,6)
d <- data.frame(time=x, score=y)
meangroup(d)

}
