% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTempSignal-package.R
\docType{package}
\name{shinyTempSignal-package}
\alias{shinyTempSignal}
\alias{shinyTempSignal-package}
\title{shinyTempSignal: Explore Temporal and Other Phylogenetic Signals}
\description{
Sequences sampled at different time points can be used to infer molecular phylogenies on natural time scales, but if the sequences records inaccurate sampling times, that are not the actual sampling times, then it will affect the molecular phylogenetic analysis. This shiny application helps exploring temporal characteristics of the evolutionary trees through linear regression analysis and with the ability to identify and remove incorrect labels. The method was extended to support exploring other phylogenetic signals under strict and relaxed models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/shinyTempSignal}
  \item \url{https://www.sciencedirect.com/science/article/pii/S167385272400033X}
  \item Report bugs at \url{https://github.com/YuLab-SMU/shinyTempSignal/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Xiao Luo \email{l77880853349@163.com} [contributor]
  \item Li Zhan \email{smu18575877413@gmail.com} [contributor]
  \item Xuanan Zhu \email{1063785553@qq.com} [contributor]
  \item Jianfeng Lin \email{415186984@qq.com} [contributor]
}

}
\keyword{internal}
