% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{shinyTempSignal_set_theme}
\alias{shinyTempSignal_set_theme}
\title{shinyTempSignal_set_theme}
\usage{
shinyTempSignal_set_theme(theme)
}
\arguments{
\item{theme}{a ggplot theme}
}
\value{
No value return
}
\description{
set default theme
}
